fix: local mode shoudn't remove cache

+++ src/common/infra/file_list/sqlite.rs
@@ -35,7 +35,7 @@ static CLIENT: Lazy<Pool<Sqlite>> = Lazy::new(connect);
 
 fn connect() -> Pool<Sqlite> {
     let url = format!("{}{}", CONFIG.common.data_cache_dir, "file_list.sqlite");
-    if std::path::Path::new(&url).exists() {
+    if !CONFIG.common.local_mode && std::path::Path::new(&url).exists() {
         std::fs::remove_file(&url).expect("remove file failed");
     }
     let db_opts = SqliteConnectOptions::from_str(&url)
